+++ src/3rd party/crypto/openssl/src/bss_file.c
@@ -194,9 +194,9 @@ static int MS_CALLBACK file_read(BIO *b, char *out, int outl)
 	if (b->init && (out != NULL))
 		{
 		if (b->flags&BIO_FLAGS_UPLINK)
-			ret=UP_fread(out,1,(int)outl,b->ptr);
+			ret=(int)UP_fread(out,1,(int)outl,b->ptr);
 		else
-			ret=fread(out,1,(int)outl,(FILE *)b->ptr);
+			ret= (int)fread(out,1,(int)outl,(FILE *)b->ptr);
 		if(ret == 0 && (b->flags&BIO_FLAGS_UPLINK)?UP_ferror((FILE *)b->ptr):ferror((FILE *)b->ptr))
 			{
 			SYSerr(SYS_F_FREAD,get_last_sys_error());
@@ -214,9 +214,9 @@ static int MS_CALLBACK file_write(BIO *b, const char *in, int inl)
 	if (b->init && (in != NULL))
 		{
 		if (b->flags&BIO_FLAGS_UPLINK)
-			ret=UP_fwrite(in,(int)inl,1,b->ptr);
+			ret= (int)UP_fwrite(in,(int)inl,1,b->ptr);
 		else
-			ret=fwrite(in,(int)inl,1,(FILE *)b->ptr);
+			ret= (int)fwrite(in,(int)inl,1,(FILE *)b->ptr);
 		if (ret)
 			ret=inl;
 		/* ret=fwrite(in,1,(int)inl,(FILE *)b->ptr); */
