+++ src/Layers/xrRender/SkeletonCustom.cpp
@@ -121,7 +121,7 @@ CKinematics::~CKinematics	()
 void	CKinematics::IBoneInstances_Create()
 {
 	// VERIFY2				(bones->size() < 64, "More than 64 bones is a crazy thing!");
-	u32				size	= bones->size();
+	u32				size	= (u32)bones->size();
 	bone_instances			=xr_alloc<CBoneInstance>(size);
 	for (u32 i=0; i<size; i++)
 		bone_instances[i].construct();
