+++ src/xrCDB/ISpatial_verify.cpp
@@ -18,7 +18,7 @@ public:
 	{
 		// test items
 		n_count			+=		1;
-		o_count			+=		N->items.size();
+		o_count			+= (u32)N->items.size();
 
 		// recurse
 		float	c_R		=		n_R/2;
