+++ src/xrGame/ui/UIChangeMap.cpp
@@ -146,7 +146,7 @@ void CUIChangeMap::FillUpList()
 	lst->Clear				();
 
 	const SGameTypeMaps& M		= gMapListHelper.GetMapListFor( (EGameIDs)GameID() );
-	u32 cnt						= M.m_map_names.size();
+	u32 cnt						= (u32)M.m_map_names.size();
 	for (u32 i=0; i<cnt; ++i)
 	{
 		CUIListBoxItem* itm		= lst->AddTextItem(g_pStringTable->translate(M.m_map_names[i].map_name).c_str() );
